Cooldown support (#266 or #454) for ApplicationCommand (SlashCommand, ContextMenuCommand(UserCommand, MessageCommand))
