+++ discord/ext/commands/errors.py
@@ -27,7 +27,7 @@ from __future__ import annotations
 
 from typing import Optional, Any, TYPE_CHECKING, List, Callable, Type, Tuple, Union
 
-from discord.errors import ClientException, DiscordException
+from discord.errors import ClientException, CommandError, CommandOnCooldown, MaxConcurrencyReached
 
 if TYPE_CHECKING:
     from inspect import Parameter
@@ -95,22 +95,6 @@ __all__ = (
     'MissingRequiredFlag',
 )
 
-class CommandError(DiscordException):
-    r"""The base exception type for all command related errors.
-
-    This inherits from :exc:`discord.DiscordException`.
-
-    This exception and exceptions inherited from it are handled
-    in a special way as they are caught and passed into a special event
-    from :class:`.Bot`\, :func:`.on_command_error`.
-    """
-    def __init__(self, message: Optional[str] = None, *args: Any) -> None:
-        if message is not None:
-            # clean-up @everyone and @here mentions
-            m = message.replace('@everyone', '@\u200beveryone').replace('@here', '@\u200bhere')
-            super().__init__(m, *args)
-        else:
-            super().__init__(*args)
 
 class ConversionError(CommandError):
     """Exception raised when a Converter class raises non-CommandError.
@@ -495,49 +479,6 @@ class CommandInvokeError(CommandError):
     def __init__(self, e: Exception) -> None:
         self.original: Exception = e
         super().__init__(f'Command raised an exception: {e.__class__.__name__}: {e}')
-
-class CommandOnCooldown(CommandError):
-    """Exception raised when the command being invoked is on cooldown.
-
-    This inherits from :exc:`CommandError`
-
-    Attributes
-    -----------
-    cooldown: :class:`.Cooldown`
-        A class with attributes ``rate`` and ``per`` similar to the
-        :func:`.cooldown` decorator.
-    type: :class:`BucketType`
-        The type associated with the cooldown.
-    retry_after: :class:`float`
-        The amount of seconds to wait before you can retry again.
-    """
-    def __init__(self, cooldown: Cooldown, retry_after: float, type: BucketType) -> None:
-        self.cooldown: Cooldown = cooldown
-        self.retry_after: float = retry_after
-        self.type: BucketType = type
-        super().__init__(f'You are on cooldown. Try again in {retry_after:.2f}s')
-
-class MaxConcurrencyReached(CommandError):
-    """Exception raised when the command being invoked has reached its maximum concurrency.
-
-    This inherits from :exc:`CommandError`.
-
-    Attributes
-    ------------
-    number: :class:`int`
-        The maximum number of concurrent invokers allowed.
-    per: :class:`.BucketType`
-        The bucket type passed to the :func:`.max_concurrency` decorator.
-    """
-
-    def __init__(self, number: int, per: BucketType) -> None:
-        self.number: int = number
-        self.per: BucketType = per
-        name = per.name
-        suffix = 'per %s' % name if per.name != 'default' else 'globally'
-        plural = '%s times %s' if number > 1 else '%s time %s'
-        fmt = plural % (number, suffix)
-        super().__init__(f'Too many people are using this command. It can only be used {fmt} concurrently.')
 
 class MissingRole(CheckFailure):
     """Exception raised when the command invoker lacks a role to run a command.
